Made some changes to tryTokenizeVariableOrKeyword() to accomodate digits when tokenizing
