+++ jamascript/src/main/java/com/jamascript/Tokenizer.java
@@ -12,20 +12,54 @@ public class Tokenizer {
         offset = 0;
     }
 
+    // Checks whether the offset is less than the string
+    // Checks also whether the current character that I'm at in the string is whitespace at the offset
     public void skipWhiteSpace() {
         while ((offset < input.length()) && Character.isWhitespace(input.charAt(offset))) {
             offset++;
         }
     }
 
+
+    public boolean isNumber(String number) {
+        boolean isANumber = true;
+        for(char c : number.toCharArray()) {
+            if(!Character.isDigit(c)) {
+                isANumber = false;
+            }
+        }
+        return isANumber;
+    }
+
     public Token tryTokenizeVariableOrKeyword() {
         skipWhiteSpace();
         String name = "";
 
+        // if the offset is less than length of string
+        // and if the current character in the string starts with a digit, it must be a digit
+        // EX: 1212321 OK
+        // EX: 1abc NOT OK
+        // EX: 12312a NOT OK
+        if(offset < input.length() && Character.isDigit(input.charAt(offset))) {
+
+            // Parse through the string until something that isn't a Letter or a Letter is encountered
+            while(Character.isDigit(input.charAt(offset)) && !Character.isWhitespace(input.charAt(offset))) {
+                name += input.charAt(offset);
+                offset++;
+            }
+
+            // might need to throw error if I encounter anything that's not a digit in this case
+
+            return new NumberToken(name);
+        }
+
+        // if the offset is less than length of string
+        // and if the current character in the string is a letter
         if((offset < input.length()) && Character.isLetter(input.charAt(offset))) {
             name += input.charAt(offset);
             offset++;
 
+            
             while((offset < input.length()) && Character.isLetterOrDigit(input.charAt(offset))) {
                 name += input.charAt(offset);
                 offset++;
@@ -39,7 +73,7 @@ public class Tokenizer {
                 return new IfToken();
             } else if(name.equals("else")) {
                 return new ElseToken();
-            } else {
+            }  else {
                 return new VariableToken(name);
             }
         } else {
